+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 ### Added
 - Added a ``create_fragment_database`` function in ``database.py`` that holds name/ql data & ratio/difference equivalents in the decitala and greek metric databases (including equivalents _across_ databases). This data is available in ``fragment_database.db`` in the databases directory. 
 
+### Changed
+- Decitala initialization relies on SQL database instead of ``os.listdir``.
+
 ## [v0.4.4] - January 10, 2020 (Kent, CT)
 ### Added
 - The output of rolling search now has an id parameter. This will be useful in a number of contexts. **NOTE**: this may require some fiddling when dealing with combined databases.